infra/src/lambda/face-crop/requirements.txt
onnxruntime==1.21.0
Pillow==11.2.1
numpy==1.26.4
boto3==1.38.0
